requirements.txt
langchain==0.1.13
pinecone-client==3.2.0
openai==1.14.2
tiktoken==0.5.2
nest-asyncio==1.5.8
sentence-transformers==2.5.1